fix(dts): preserve literal short-circuit unions
