+++ crates/tsz-emitter/src/declaration_emitter/tests/mod.rs
@@ -296,6 +296,7 @@ mod computed_properties;
 mod enum_template_and_advanced;
 mod export_specifiers;
 mod fix_verification;
+mod fix_verification_returned;
 mod generics_and_ambient;
 mod js_define_property;
 mod jsdoc_template_defaults;
